+++ README.md
@@ -1,5 +1,5 @@
-# deep-metadetection
+# deep-field-metadetect
 
-[![tests](https://github.com/beckermr/deep-metadetection/actions/workflows/tests.yml/badge.svg)](https://github.com/beckermr/deep-metadetection/actions/workflows/tests.yml) [![pre-commit.ci status](https://results.pre-commit.ci/badge/github/beckermr/deep-metadetection/main.svg)](https://results.pre-commit.ci/latest/github/beckermr/deep-metadetection/main) [![codecov](https://codecov.io/gh/beckermr/deep-metadetection/graph/badge.svg?token=CHqIu9leXU)](https://codecov.io/gh/beckermr/deep-metadetection)
+[![tests](https://github.com/beckermr/deep-field-metadetect/actions/workflows/tests.yml/badge.svg)](https://github.com/beckermr/deep-field-metadetect/actions/workflows/tests.yml) [![pre-commit.ci status](https://results.pre-commit.ci/badge/github/beckermr/deep-field-metadetect/main.svg)](https://results.pre-commit.ci/latest/github/beckermr/deep-field-metadetect/main) [![codecov](https://codecov.io/gh/beckermr/deep-field-metadetect/graph/badge.svg?token=CHqIu9leXU)](https://codecov.io/gh/beckermr/deep-field-metadetect)
 
 deep-field metadetection
